chore: move handler to new file
